LayoutTests/media/modern-media-controls/controls-visibility-support/controls-visibility-support-fullscreen-on-video-expected.txt

Testing that control are shown in fullscreen when the controls attribute is not present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isChangingPresentationMode(media) became false
PASS mediaController.controls.visible is true
PASS successfullyParsed is true

TEST COMPLETE
